+++ gulpfile.js
@@ -1,14 +1,17 @@
 /*
  * GULP CONFIG
  *
- * Desciption:  Clean gulpfile for web development workflow using
+ * Desciption:  Clean gulpfile for web development workflow containing
+ *              - compiling/optimization of sass, javascript and images from assets to dist and vendors
  *              - browsersync
- *              - compiling/optimization of sass, javascript and images from assets to dist and vendors (usally from npm_modules to dist)
- * Usage:       gulp (to run the whole process), gulp watch (to watch for changes and compile if anything is being modified)
+ *              - cache-busting
+ *
+ * Usage:       - gulp (to run the whole process)
+ *              - gulp watch (to watch for changes and compile if anything is being modified)
  *
  * Author:      Flurin Dürst (https://flurinduerst.ch)
  *
- * Version:     1.4.0
+ * Version:     2.0.0
  *
 */
 
@@ -25,10 +28,13 @@ var gulp = require('gulp');
 var browserSync = require('browser-sync').create();
 var concat = require('gulp-concat');
 var rename = require("gulp-rename");
+// var clean = require('gulp-clean');
 // css
 var sass = require('gulp-sass');
 var cleanCSS = require('gulp-clean-css');
 var autoprefixer = require('gulp-autoprefixer');
+// cache busting
+var rev = require('gulp-rev');
 // js
 var uglify = require('gulp-uglify');
 // images
@@ -39,11 +45,32 @@ var plumber = require('gulp-plumber');
 // watch
 var watch = require('gulp-watch');
 // delete
-// var del = require('del');
+var del = require('del');
 
 
 /* TASKS
 /===================================================== */
+
+/* CLEAN
+/------------------------*/
+// delete compiled files/folders (before running the build)
+// css
+gulp.task('clean:css', function() {
+  return del(['dist/*.css', 'dist/rev-manifest.json']);
+});
+// cachebust
+gulp.task('clean:cachebust', function() {
+  return del(['dist/style-*.min.css']);
+});
+// javascript
+gulp.task('clean:javascript', function() {
+  return del(['dist/*.js']);
+});
+// images
+gulp.task('clean:images', function() {
+  return del(['dist/images']);
+});
+
 
 /* BROWSERSYNC
 /------------------------*/
@@ -60,13 +87,14 @@ gulp.task('browsersync', function() {
   });
 });
 
+
 /* CSS
 /------------------------*/
 // from:    assets/styles/main.css (+ optional vendors)
 // actions: compile, minify, prefix, rename
 // to:      dist/style.min.css
-gulp.task('css', function() {
-  gulp.src([
+gulp.task('css', ['clean:css'], function() {
+  return gulp.src([
     //main
     'assets/styles/bundle.scss',
     // vendors
@@ -83,13 +111,28 @@ gulp.task('css', function() {
   .pipe(browserSync.stream());
 });
 
+
+/* CSS CACHE BUSTING
+/------------------------*/
+// from:    dist/style.min.css
+// actions: create busted version of file
+// to:      dist/style-[hash].min.css
+gulp.task('cachebust', ['clean:cachebust', 'css'], function() {
+  return gulp.src('dist/style.min.css')
+    .pipe(rev())
+    .pipe(gulp.dest('dist'))
+    .pipe(rev.manifest({merge: true}))
+    .pipe(gulp.dest('dist'))
+});
+
+
 /* JAVASCRIPT
 /------------------------*/
 // from:    assets/scripts/ (+ optional vendors)
 // actions: concatinate, minify, rename
 // to:      dist/script.min.css
-gulp.task('javascript', function() {
-  gulp.src([
+gulp.task('javascript', ['clean:javascript'], function() {
+  return gulp.src([
     // main
     'assets/scripts/*.js',
     // vendors
@@ -103,36 +146,34 @@ gulp.task('javascript', function() {
   .pipe(browserSync.stream());
 });
 
+
 /* IMAGES
 /------------------------*/
 // from:    assets/images/
 // actions: minify
 // to:      dist/images
-gulp.task('images',  function() {
-  gulp.src('assets/images/*.*')
+gulp.task('images', ['clean:images'],  function() {
+  return gulp.src('assets/images/*.*')
     .pipe(imagemin())
     .pipe(gulp.dest('dist/images'))
     // .pipe(browserSync.stream()); // currently bugged (18.12.2017)
 });
 
-/* CLEAN
-/------------------------*/
-// empty dist folder
-gulp.task('clean', require('del').bind(null, ['dist/*']));
 
 /* WATCH
 /------------------------*/
 // watch for modifications in
 // styles, scripts, images, php files, html files
 gulp.task('watch',  ['browsersync'], function() {
-  gulp.watch('assets/styles/*.scss', ['css']);
+  gulp.watch('assets/styles/*.scss', ['css', 'cachebust']);
   gulp.watch('assets/scripts/*.js', ['javascript']);
   gulp.watch('assets/images/*.*', ['images']);
   gulp.watch('*.php', browserSync.reload);
   gulp.watch('*.html', browserSync.reload);
 });
 
+
 /* DEFAULT
 /------------------------*/
 // default gulp tasks executed with `gulp`
-gulp.task('default', ['clean', 'css', 'javascript', 'images']);
+gulp.task('default', ['css', 'cachebust', 'javascript', 'images']);